requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pypdf>=3.17.1
openai>=1.3.0
numpy==1.24.3
pandas==2.0.3 